+++ backend/framework/sdk/src/main/java/io/metersphere/sdk/mapper/BaseMapper.java
@@ -1,0 +1,4 @@
+package io.metersphere.sdk.mapper;
+
+public interface BaseMapper {
+}
